Update index.html
Setup for the time being, I would like to remove the javascript and make the site no-script friendly but for now learning the CSS would be less than productive for this single use case. 

I have commented out elements that I plan to use in the future but for now I am happy with the way it looks and it allows navigation from my homepage to keybase and my blog.